Create class for create objects of the class Chat
This class loading list of Chat objects for initializing for the first time. Then it add listeners for classes which want to listen changed in loaded list (adding new object or deleting them)

All dates for Chat object gets from string resources so this class get Context in own constructor. Only addresses of images we safe in this class.